Separate the proxies startup to their own function
